lms-pay-clean/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lms-pay"
version = "0.1.0"
description = "Payment service for LMS"
authors = [
    { name ="Ng Cao Thanh", email = "cthanhnguyen03@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
    "sqlalchemy[asyncio]",
    "psycopg2-binary",
    "python-dotenv",
    "pydantic",
    "grpcio-reflection",
    "asyncio",
    "asyncpg",
    "uvicorn[standard]",
    "fastapi",
    "passlib[bcrypt]",
    "bcrypt==3.2.0",
    "pyjwt",
    "aio_pika",
    "injector",
    "fastapi_mail",
    "python-multipart",
    "pandas",
    "openpyxl",
]

[tool.setuptools]
packages = ["app", "alembic"]